requirements.txt
streamlit==1.22.0
folium==0.14.0
streamlit-folium==0.11.0
pandas==1.5.3
plotly==5.14.1
uuid==1.30
python-dotenv==1.0.0
aixplain
twilio==7.16.0
requests==2.28.2
sqlite3